Removed pupil prefs storage from aa-fonts component
- Moved pupil prefs storage code from aa-fonts component into the pupil prefs service
- Added pupil prefs service to aa-fonts and aa-colours
